requirements.txt
openai>=1.0.0
anthropic>=0.8.0
mistralai>=0.0.7
pyyaml>=6.0
requests>=2.30.0
